fix(higgs-tts): patch null rope_theta + cast fused modules to bf16 — seed-tts parity with boson-vllm (#15)
* wip(bench): boson-vllm vs sglang-omni parity benchmarks (seed-tts WER)

Tooling to measure the quality gap between our Higgs TTS pipeline and
boson-vllm 0.14's reference serving of the same checkpoint. Findings
captured in ``bench/README.md``; TL;DR is we have a real and large
quality gap (10-sample average WER 0.586 vs boson's 0.017 at
temperature 0.8 / top_k 50) that the quick-audit paths below don't
close.

Scripts:
- ``compare_boson_vs_sglang.py`` — E2E WER + code-length comparison.
- ``diff_codes.py`` — greedy per-step code diff (confirms row 0-1 match, row 2 diverges).
- ``inspect_cb0_eoc.py`` — dumps top-10 cb0 + EOC probability to rule out premature EOC.
- ``verify_overlay.py`` — confirms our ``_inject_ref_audio_prefill`` overlay reaches the
  backbone (input_embeds norms match between inject-site and backbone.forward call).
- ``README.md`` — results + things-tried + follow-up paths.

Not landing code changes yet — the bug isn't in the surface-level glue
(prompt template, sampler math, gen_params threading, overlay plumbing
all verified correct). Next step: either replace the sglang Qwen3
backbone with HF transformers' ``Qwen3Model`` (voxtral-style AR) to
isolate whether the sglang Qwen3 kernel path is the cause, or diff
intermediate hidden states between stacks.

* fix(higgs-tts): patch null rope_theta + cast fused modules to backbone dtype — matches boson-vllm on seed-tts

Higgs TTS checkpoints ship ``text_config.rope_theta: null``. When realised
via ``Qwen3Config(**dict)``, transformers silently falls back to its Qwen3
default of 10000 — but Higgs TTS was trained with ``rope_theta=1e6`` (the
Qwen3 convention). boson-vllm's model init calls
``set_default_rope_theta(config, default_theta=1000000)`` up front; sglang's
Qwen3 reads ``getattr(config, "rope_theta", 1000000)``, which does NOT
trigger on explicit ``None``. Result: our backbone ran with the wrong
rotary base, silently wrecking positional encoding.

Symptom: seed-tts-eval English WER 0.586 vs boson-vllm's 0.017 at matched
temperature=0.8 / top_k=50, and greedy codes diverge between the two stacks
starting row 2 despite matching rows 0-1.

Fix: in ``HiggsMultimodalQwen3Config.__init__`` and ``.get_text_config()``,
rewrite ``rope_theta=None`` to ``1_000_000`` before constructing
``Qwen3Config``. Regression test asserts ``get_text_config().rope_theta ==
1_000_000`` when the incoming dict has a null ``rope_theta``.

Two 1-ULP bf16 drifts surfaced while chasing this are also fixed:

- ``_load_fused_embedding_from_tts_ckpt`` in ``pipeline/stages.py`` now
  casts the fused embedding module to bf16 BEFORE copying the weight, so
  the audio_encoder stage's forward (lookup + sum-over-codebooks) runs in
  bf16 and matches boson-vllm's inline ``embed_multimodal`` path bitwise.
- ``HiggsTTSModel.__init__`` now casts ``multimodal_embedding`` and
  ``modality_head`` to the backbone's dtype. Without this the
  default-initialised ``nn.Parameter`` stayed in fp32, and every
  decode-step re-embed accumulated 1 bf16 ULP of drift vs boson.

Validation:
- ``bench/diff_codes.py`` (greedy, temperature=0, top_k=1): rows 0-2 now
  match boson-vllm byte-for-byte (previously row 2 diverged).
- ``bench/compare_boson_vs_sglang.py`` (temperature=0.8, top_k=50,
  seed-tts-eval EN): 100-sample average WER 0.0559 ours vs 0.0638 boson —
  parity within Whisper + multinomial-sampling noise (previously 0.586 vs
  0.017). Median WER 0 on both sides.

Also adds ``bench/diff_step1_prompt.py`` and ``bench/diff_step3_fused_embed.py``
(used while diagnosing this) and updates ``bench/README.md`` with the new
results and root-cause writeup.


---------

+++ sglang_omni/models/higgs_tts/pipeline/stages.py
@@ -340,8 +340,15 @@ def _load_fused_embedding_from_tts_ckpt(tts_ckpt_path: str, device: str = "cpu")
             f"Fused embedding shape mismatch: ckpt {tuple(tensor.shape)}, "
             f"expected {tuple(module.weight.shape)}"
         )
+    # Cast the module to bfloat16 BEFORE copying the weight, so forward
+    # compute (lookup + sum-over-codebooks) runs in bf16 — matching the
+    # engine backbone's dtype and boson-vllm's inline embed_multimodal
+    # path. If we leave the module in fp32, the summed embedding ends
+    # up with ~1 bf16 ULP more precision than boson produces, which
+    # compounds over the AR loop.
+    module = module.to(dtype=torch.bfloat16)
     with torch.no_grad():
-        module.weight.copy_(tensor)
+        module.weight.copy_(tensor.to(dtype=torch.bfloat16))
 
     module = module.to(device).eval()
     for p in module.parameters():
